requirements.txt
kafka-python==2.3.0
pytest==9.0.2
